core: fix bug in Helgrind header
Current code in upstream header generates a sign-conversion error
and actually returns UINT32_MAX-1 when not running under Helgrind.

+++ src/core/valgrind/helgrind.h
@@ -218,7 +218,8 @@ typedef
       _arg1 = (long int)(_arg1F);                        \
       _arg2 = (long int)(_arg2F);                        \
       _arg3 = (long int)(_arg3F);                        \
-      _qzz_res = VALGRIND_DO_CLIENT_REQUEST_EXPR(        \
+      /* XXX: here PMDK's version deviate from upstream */\
+      _qzz_res = (long)VALGRIND_DO_CLIENT_REQUEST_EXPR(  \
                                  (_dfltF),               \
                                  (_creqF),               \
                                  _arg1,_arg2,_arg3,0,0); \
@@ -418,7 +419,8 @@ typedef
 #define VALGRIND_HG_GET_ABITS(zza,zzabits,zznbytes)          \
    (__extension__                                            \
    ({long int _res;                                          \
-      DO_CREQ_W_WWW(_res, (-2)/*default*/,                   \
+      /* XXX: here PMDK's version deviate from upstream */   \
+      DO_CREQ_W_WWW(_res, (-2LL)/*default*/,                 \
                     _VG_USERREQ__HG_GET_ABITS,               \
                     void*,(zza), void*,(zzabits),            \
                     unsigned long,(zznbytes));               \
